# Conflicts:
#	src/main/java/com/genersoft/iot/vmp/gb28181/transmit/event/request/impl/message/response/cmd/RecordInfoResponseMessageHandler.java
#	src/main/resources/application-dev.yml
#	src/main/resources/application.yml

+++ src/main/java/com/genersoft/iot/vmp/gb28181/service/impl/PlatformChannelServiceImpl.java
@@ -1,6 +1,5 @@
 package com.genersoft.iot.vmp.gb28181.service.impl;
 
-import com.baomidou.dynamic.datasource.annotation.DS;
 import com.genersoft.iot.vmp.common.enums.ChannelDataType;
 import com.genersoft.iot.vmp.gb28181.bean.*;
 import com.genersoft.iot.vmp.gb28181.dao.*;
@@ -26,7 +25,6 @@ import java.util.*;
  */
 @Slf4j
 @Service
-@DS("master")
 public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Autowired
@@ -215,6 +213,10 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Transactional
     public int addChannelList(Integer platformId, List<CommonGBChannel> channelList) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
         int result = platformChannelMapper.addChannels(platformId, channelList);
         if (result > 0) {
             // 查询通道相关的行政区划信息是否共享，如果没共享就添加
@@ -244,7 +246,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelList, CatalogEvent.ADD);
+                eventPublisher.catalogEventPublish(platform, channelList, CatalogEvent.ADD);
             } catch (Exception e) {
                 log.warn("[关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -254,6 +256,11 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Override
     public int removeAllChannel(Integer platformId) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
+
         List<CommonGBChannel> channelListShare = platformChannelMapper.queryShare(platformId,  null);
         Assert.notEmpty(channelListShare, "未共享任何通道");
         int result = platformChannelMapper.removeChannelsWithPlatform(platformId, channelListShare);
@@ -278,7 +285,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelListShare, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListShare, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除全部关联通道] 发送失败，数量：{}", channelListShare.size(), e);
             }
@@ -302,6 +309,10 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
 
     @Transactional
     public int removeChannelList(Integer platformId, List<CommonGBChannel> channelList) {
+        Platform platform = platformMapper.query(platformId);
+        if (platform == null) {
+            return 0;
+        }
         int result = platformChannelMapper.removeChannelsWithPlatform(platformId, channelList);
         if (result > 0) {
             // 查询通道相关的分组信息
@@ -324,7 +335,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platformId, channelList, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelList, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -425,11 +436,12 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
     @Override
     public void updateCustomChannel(PlatformChannel channel) {
         platformChannelMapper.updateCustomChannel(channel);
+        Platform platform = platformMapper.query(channel.getPlatformId());
         CommonGBChannel commonGBChannel = platformChannelMapper.queryShareChannel(channel.getPlatformId(), channel.getGbId());
         // 发送消息
         try {
             // 发送catalog
-            eventPublisher.catalogEventPublish(channel.getPlatformId(), commonGBChannel, CatalogEvent.UPDATE);
+            eventPublisher.catalogEventPublish(platform, commonGBChannel, CatalogEvent.UPDATE);
         } catch (Exception e) {
             log.warn("[自定义通道信息] 发送失败， 平台ID： {}， 通道： {}（{}）", channel.getPlatformId(),
                     channel.getGbName(), channel.getId(), e);
@@ -468,7 +480,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -506,7 +518,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
             // 发送消息
             try {
                 // 发送catalog
-                eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.DEL);
+                eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.DEL);
             } catch (Exception e) {
                 log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
             }
@@ -537,7 +549,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
                 // 发送消息
                 try {
                     // 发送catalog
-                    eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.ADD);
+                    eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.ADD);
                 } catch (Exception e) {
                     log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
                 }
@@ -567,7 +579,7 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
                 // 发送消息
                 try {
                     // 发送catalog
-                    eventPublisher.catalogEventPublish(platform.getId(), channelListForEvent, CatalogEvent.ADD);
+                    eventPublisher.catalogEventPublish(platform, channelListForEvent, CatalogEvent.ADD);
                 } catch (Exception e) {
                     log.warn("[移除关联通道] 发送失败，数量：{}", channelList.size(), e);
                 }
@@ -583,5 +595,18 @@ public class PlatformChannelServiceImpl implements IPlatformChannelService {
     @Override
     public CommonGBChannel queryChannelByPlatformIdAndChannelId(Integer platformId, Integer channelId) {
         return platformChannelMapper.queryShareChannel(platformId, channelId);
+    }
+
+    @Override
+    public List<CommonGBChannel> queryChannelByPlatformIdAndChannelIds(Integer platformId, List<Integer> channelIds) {
+        return platformChannelMapper.queryShare(platformId, channelIds);
+    }
+
+    @Override
+    public List<Platform> queryByPlatformBySharChannelId(String channelDeviceId) {
+        CommonGBChannel commonGBChannel = commonGBChannelMapper.queryByDeviceId(channelDeviceId);
+        ArrayList<Integer> ids = new ArrayList<>();
+        ids.add(commonGBChannel.getGbId());
+        return platformChannelMapper.queryPlatFormListByChannelList(ids);
     }
 }
